meta: Attempt to convert to manifest v3
This was my attempt at converting to manifest v3. Note that this change
requires users to manually open about:addons, go to the permissions tab,
and enable "Access your data for all websites", which is a lot of work.

Forcing the user to manually enable permissions when they previously
didn't have to is additionally a lot of work, especially when the addon
previously worked out of the box with no intervention necessary.

References:

- https://stackoverflow.com/questions/66665810/cors-and-content-policy-violation-problems-chrome-extension-manifest-v3/66666711#66666711
- https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API/runtime/onMessage
- https://stackoverflow.com/questions/55214828/how-to-make-a-cross-origin-request-in-a-content-script-currently-blocked-by-cor/55292071#55292071
- https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API/permissions/request
- https://extensionworkshop.com/documentation/develop/manifest-v3-migration-guide/